CalendarApp added.
Appointment is not added yet.
I didn't add any image for design for now. Because we've got project size limit. So i dont wanna pass that limit without finishing all the features that i want it to be in.
There is one little incomplete code about File I/O but im gonna fix it soon.
All other features are available.