Applications can be added dynamicly.
-ApplicationManager extends Properties. Properties are used to load
and store Application informations thar locates in user.home/yna.
ApplicatipnManager not only loads and stores, it also gives a way
to Applications with given 'extension'-getApplications(String).

createAssociation(App) if extensions not created before, it creates
the extension and matches it with the app. If exist it just updates
with new app.

-Each Applicaton needs to extends SimpleApplication to use in
browser. It basicly a JFrame. SimpleApplication have methods to
waits implement from subclass. Like run(), this needs to create a
instance by Reflection.
PaintApplicaiton extends it and provides how should run application.

-Reflect.java modified. getClass is overloaded.

-Browser.java is not ready for now. It is on progress. If i could
mention about it, it will read all applications and create-update
associations for current session.